Add abstract options.Driver, use for Driver
This creates an abstract class representing a Big5 Driver and uses it as
the parent class for FIRRTL's Driver.
