+++ src/main/scala/firrtl/options/Driver.scala
@@ -1,0 +1,40 @@
+// See LICENSE for license details.
+
+package firrtl.options
+
+import firrtl.AnnotationSeq
+
+trait DriverExecutionResult
+
+abstract class Driver {
+  val optionsManager: ExecutionOptionsManager
+
+  def execute(args: Array[String], initialAnnotations: AnnotationSeq = Seq.empty): DriverExecutionResult
+
+  def main(args: Array[String]): Unit = execute(args)
+
+  /**
+    * Print a warning message (in yellow)
+    *
+    * @param message error message
+    */
+  //scalastyle:off regex
+  def dramaticWarning(message: String): Unit = {
+    println(Console.YELLOW + "-"*78)
+    println(s"Warning: $message")
+    println("-"*78 + Console.RESET)
+  }
+
+  /**
+    * Print an error message (in red)
+    *
+    * @param message error message
+    * @note This does not stop the Driver.
+    */
+  //scalastyle:off regex
+  def dramaticError(message: String): Unit = {
+    println(Console.RED + "-"*78)
+    println(s"Error: $message")
+    println("-"*78 + Console.RESET)
+  }
+}
